requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
python-decouple>=3.8,<4.0
django-cors-headers>=4.3,<5.0
